speedup and refactor

+++ test/unit/cache/triggers/one-row/has_cost_sale/operation.operation.sql
@@ -9,15 +9,15 @@ begin
             old.deleted = 0) is not null
         then
             update operation.fin_operation as fin_op set
-                is_cost_sale = (null is not null
+                is_cost_sale = ((null::numeric) is not null
                 and
-                null = 0)
+                (null::smallint) = 0)
             where
                 old.id = fin_op.id_source_operation_sale
                 and
-                fin_op.is_cost_sale is distinct from ((null is not null
+                fin_op.is_cost_sale is distinct from (((null::numeric) is not null
                 and
-                null = 0));
+                (null::smallint) = 0));
         end if;
 
         return old;
